+++ CONTRIBUTORS.md
@@ -3,3 +3,4 @@
 | GitHub user | Real Name | Affiliation | Date |
 | ----------- | --------- | ----------- | ---- |
 | james-bruten-mo | James Bruten | Met Office | 2025-12-09 |
+| jennyhickson | Jenny Hickson | Met Office | 2025-12-10 |
