+++ applications/io_demo/rose-meta/lfric-io_demo/versions.py
@@ -2,7 +2,7 @@ import sys
 
 from metomi.rose.upgrade import MacroUpgrade
 
-from .version21_22 import *
+from .version22_30 import *
 
 
 class UpgradeError(Exception):
@@ -31,32 +31,3 @@ class vnXX_txxx(MacroUpgrade):
         # Add settings
         return config, self.reports
 """
-
-
-class vn22_t4661(MacroUpgrade):
-    """Upgrade macro for ticket #4661 by Denis Sergeev."""
-
-    BEFORE_TAG = "vn2.2"
-    AFTER_TAG = "vn2.2_t4661"
-
-    def upgrade(self, config, meta_config=None):
-        # Commands From: rose-meta/lfric-driver
-        self.add_setting(config, ["namelist:extrusion", "eta_values"], "''")
-
-        return config, self.reports
-
-
-class vn22_t4020(MacroUpgrade):
-    """Upgrade macro for ticket #4020 by Andrew Coughtrie."""
-
-    BEFORE_TAG = "vn2.2_t4661"
-    AFTER_TAG = "vn2.2_t4020"
-
-    def upgrade(self, config, meta_config=None):
-        # Commands From: rose-meta/lfric-driver
-        self.add_setting(
-            config, ["namelist:io", "end_of_run_checkpoint"], ".true."
-        )
-        self.add_setting(config, ["namelist:io", "checkpoint_times"], "")
-
-        return config, self.reports
