+++ applications/io_demo/rose-meta/lfric-io_demo/vn3.0/rose-meta.conf
@@ -1,0 +1,29 @@
+import=lfric-driver/vn3.0
+
+[namelist:io=multifile_io]
+compulsory=true
+description=Use multifile_io functionality
+help=This is used to turn the multifile_io functionality in the io_demo app
+    =on and off
+!kind=default
+type=logical
+
+[namelist:multifile_io]
+compulsory=false
+duplicate=true
+!instance_key_member=filename
+
+[namelist:multifile_io=filename]
+compulsory=true
+!kind=default
+type=character
+
+[namelist:multifile_io=start_timestep]
+compulsory=true
+!kind=default
+type=integer
+
+[namelist:multifile_io=stop_timestep]
+compulsory=true
+!kind=default
+type=integer
