+++ applications/lbc_demo/rose-meta/lfric-lbc_demo/versions.py
@@ -1,5 +1,9 @@
 import sys
+
 from metomi.rose.upgrade import MacroUpgrade
+
+from .version22_30 import *
+
 
 class UpgradeError(Exception):
     """Exception created when an upgrade fails."""
@@ -13,6 +17,7 @@ class UpgradeError(Exception):
 
     __str__ = __repr__
 
+
 """
 Copy this template and complete to add your macro
 
@@ -21,20 +26,6 @@ class vnXX_txxx(MacroUpgrade):
 
     BEFORE_TAG = "vnX.X"
     AFTER_TAG = "vnX.X_txxx"
-
-    def upgrade(self, config, meta_config=None):
-        # Add settings
-        return config, self.reports
-"""
-
-class vn22_t4231(MacroUpgrade):
-    # Upgrade macro for #4231 by Ricky Wong
-    # Ticket #4231 adds a new application, "lbc_demo"
-    # This ugrade macro does nothing but set the
-    # appropriate "After_tag"
-
-    BEFORE_TAG = "vn2.2"
-    AFTER_TAG = "vn2.2_t4231"
 
     def upgrade(self, config, meta_config=None):
         # Add settings
