+++ applications/lbc_demo/rose-meta/lfric-lbc_demo/vn3.0/rose-meta.conf
@@ -1,0 +1,42 @@
+import=lfric-driver/vn3.0
+
+[namelist:lbc_demo]
+compulsory=true
+
+[namelist:lbc_demo=apply_lbc]
+compulsory=true
+!kind=default
+type=logical
+
+[namelist:lbc_demo=enable_lbc]
+compulsory=true
+!kind=default
+type=logical
+
+[namelist:lbc_demo=field_type]
+compulsory=true
+!enumeration=true
+value-titles='Real','Integer'
+values='real','integer'
+
+[namelist:lbc_demo=lbc_source]
+compulsory=true
+!enumeration=true
+value-titles='File','Analytic'
+values='file','analytic'
+
+[namelist:lbc_demo=read_lbc]
+compulsory=true
+!kind=default
+type=logical
+
+[namelist:lbc_demo=set_lbc]
+compulsory=true
+!enumeration=true
+value-titles='Constant','Quadrant'
+values='constant','quadrant'
+
+[namelist:lbc_demo=write_lbc]
+compulsory=true
+!kind=default
+type=logical
