+++ applications/skeleton/rose-meta/lfric-skeleton/vn3.0/rose-meta.conf
@@ -1,0 +1,1 @@
+import=lfric-driver/vn3.0
