+++ dependencies.yaml
@@ -23,4 +23,4 @@ lfric_core:
 
 SimSys_Scripts:
     source: git@github.com:MetOffice/SimSys_Scripts.git
-    ref: main
+    ref: 2025.12.1
