+++ rose-stem/app/io_demo/rose-app.conf
@@ -1,4 +1,4 @@
-meta=lfric-io_demo/vn2.2_t4020
+meta=lfric-io_demo/vn3.0
 
 [command]
 default=$CORE_ROOT_DIR/bin/tweak_iodef ; \
