+++ rose-stem/app/mesh/opt/rose-app-lbc_8x2P.conf
@@ -3,19 +3,13 @@ default=mkdir -p $MESH_DIR ; mpiexec -n 1 $BIN_DIR/planar_mesh_generator mesh_ge
 srun=mkdir -p $MESH_DIR ; srun --ntasks=1 $BIN_DIR/planar_mesh_generator mesh_generation.nml
 
 [env]
-OMP_NUM_THREADS=1
 mesh_generator=Planar
 
 [!!namelist:cubedsphere_mesh]
 
 [namelist:mesh]
-coord_sys='ll'
-geometry='spherical'
-mesh_file_prefix='$OUTPUT_FILE_PREFIX'
 mesh_names='primary'
-n_meshes=1
 partition_mesh=.true.
-rotate_mesh=.false.
 topology='non_periodic'
 
 [namelist:partitions]
@@ -27,9 +21,7 @@ panel_yproc=2
 partition_range=0,15
 
 [namelist:planar_mesh]
-apply_stretch_transform=.false.
 create_lbc_mesh=.true.
-domain_centre=0.0,0.0
 domain_size=24.0,22.0
 edge_cells_x=96
 edge_cells_y=96
@@ -37,7 +29,3 @@ lbc_parent_mesh='primary'
 lbc_rim_depth=20
 periodic_x=.false.
 periodic_y=.false.
-
-[!!namelist:rotation]
-
-[!!namelist:stretch_transform]
