+++ rose-stem/app/mesh/rose-app.conf
@@ -1,4 +1,4 @@
-meta=lfric-mesh_tools/vn2.2
+meta=lfric-mesh_tools/vn3.0
 
 [command]
 default=echo "There is no default mesh generator, please specify an optional configuration"; false
