+++ rose-stem/app/mesh_tools/rose-app.conf
@@ -1,4 +1,4 @@
-meta=lfric-mesh_tools/vn2.2
+meta=lfric-mesh_tools/vn3.0
 
 [command]
 default=${BIN_DIR}/${APPLICATION} mesh_generation.nml
