+++ rose-stem/app/simple_diffusion/rose-app.conf
@@ -1,4 +1,4 @@
-meta=lfric-simple_diffusion/vn2.2_t4020
+meta=lfric-simple_diffusion/vn3.0
 
 [command]
 default=$CORE_ROOT_DIR/bin/tweak_iodef ; \
