+++ rose-stem/app/skeleton/rose-app.conf
@@ -1,4 +1,4 @@
-meta=lfric-skeleton/vn2.2_t4020
+meta=lfric-skeleton/vn3.0
 
 [command]
 default=$LAUNCH_SCRIPT/launch-exe
